# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NER Smart Logistics & Accessibility Intelligence Platform",
    "description": "Backend API for the Smart Logistics & Accessibility Intelligence Platform for North-East India (SIH 2026 \u2014 Problem 26002, MDoNER). Supports field incident reporting, road network status, routing, fleet tracking, and logistics coordination for humanitarian operations.",
    "contact": {
      "name": "SIH 2026 Team"
    },
    "license": {
      "name": "Restricted \u2014 Pilot Use Only"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/oidc/callback": {
      "get": {
        "tags": [
          "Identity & Auth"
        ],
        "summary": "Oidc Callback",
        "description": "Process OIDC provider redirect callback.",
        "operationId": "oidc_callback_api_v1_auth_oidc_callback_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Oidc Callback Api V1 Auth Oidc Callback Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/select-org": {
      "post": {
        "tags": [
          "Identity & Auth"
        ],
        "summary": "Select Org",
        "description": "Select active organization context for multi-org users.",
        "operationId": "select_org_api_v1_auth_select_org_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectOrgRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Select Org Api V1 Auth Select Org Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/dev-session": {
      "post": {
        "tags": [
          "Identity & Auth"
        ],
        "summary": "Dev Session",
        "description": "DEV ONLY: Create a full server session for local development without OIDC.\nGuarded by DevModeVerifier \u2014 strictly blocked in staging and production.",
        "operationId": "dev_session_api_v1_auth_dev_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Dev Session Api V1 Auth Dev Session Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/routes/{route_plan_id}": {
      "get": {
        "tags": [
          "Routing & Dispatch"
        ],
        "summary": "Get Route Plan",
        "operationId": "get_route_plan_api_v1_routes__route_plan_id__get",
        "parameters": [
          {
            "name": "route_plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Route Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutePlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlternativeRouteResponse": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "total_distance_meters": {
            "type": "integer",
            "title": "Total Distance Meters"
          },
          "total_duration_seconds": {
            "type": "integer",
            "title": "Total Duration Seconds"
          },
          "geometry": {
            "type": "object",
            "title": "Geometry"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/RouteEdgeResponse"
            },
            "type": "array",
            "title": "Edges"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "total_distance_meters",
          "total_duration_seconds",
          "geometry",
          "edges"
        ],
        "title": "AlternativeRouteResponse"
      },
...
      "DevSessionRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "org_id": {
            "type": "string",
            "format": "uuid",
            "title": "Org Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "org_id",
          "role"
        ],
        "title": "DevSessionRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RouteEdgeResponse": {
        "properties": {
          "edge_id": {
            "type": "string",
            "format": "uuid",
            "title": "Edge Id"
          },
          "sequence_order": {
            "type": "integer",
            "title": "Sequence Order"
          },
          "cumulative_distance_meters": {
            "type": "integer",
            "title": "Cumulative Distance Meters"
          },
          "cumulative_duration_seconds": {
            "type": "integer",
            "title": "Cumulative Duration Seconds"
          }
        },
        "type": "object",
        "required": [
          "edge_id",
          "sequence_order",
          "cumulative_distance_meters",
          "cumulative_duration_seconds"
        ],
        "title": "RouteEdgeResponse"
      },
...
      "RoutePlanResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "title": "Organization Id"
          },
          "trip_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trip Id"
          },
          "graph_version": {
            "type": "string",
            "title": "Graph Version"
          },
          "status_version": {
            "type": "integer",
            "title": "Status Version"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version"
          },
          "result_status": {
            "$ref": "#/components/schemas/RouteResultStatus"
          },
          "total_distance_meters": {
            "type": "integer",
            "title": "Total Distance Meters"
          },
          "total_duration_seconds": {
            "type": "integer",
            "title": "Total Duration Seconds"
          },
          "requires_human_review": {
            "type": "boolean",
            "title": "Requires Human Review"
          },
          "excluded_edge_reasons": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Excluded Edge Reasons"
          },
          "primary_geometry": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Geometry"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/RouteEdgeResponse"
            },
            "type": "array",
            "title": "Edges"
          },
          "alternatives": {
            "items": {
              "$ref": "#/components/schemas/AlternativeRouteResponse"
            },
            "type": "array",
            "title": "Alternatives"
          },
          "evaluated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Evaluated At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "trip_id",
          "graph_version",
          "status_version",
          "policy_version",
          "result_status",
          "total_distance_meters",
          "total_duration_seconds",
          "requires_human_review",
          "excluded_edge_reasons",
          "primary_geometry",
          "edges",
          "alternatives",
          "evaluated_at",
          "expires_at"
        ],
        "title": "RoutePlanResponse"
      },
      "RouteResultStatus": {
        "type": "string",
        "enum": [
          "FEASIBLE",
          "NO_FEASIBLE_PATH",
          "INSUFFICIENT_DATA"
        ],
        "title": "RouteResultStatus"
      },
      "SelectOrgRequest": {
        "properties": {
          "org_id": {
            "type": "string",
            "format": "uuid",
            "title": "Org Id"
          },
          "selection_token": {
            "type": "string",
            "title": "Selection Token"
          }
        },
        "type": "object",
        "required": [
          "org_id",
          "selection_token"
        ],
        "title": "SelectOrgRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}